moved index&fieldspec logic from spec.h to field_spec_info.h
