+++ src/spec.h
@@ -18,7 +18,7 @@
 #include "gc.h"
 #include "synonym_map.h"
 #include "query_error.h"
-#include "info/field_spec_info.h"
+#include "field_spec.h"
 #include "util/dict.h"
 #include "util/references.h"
 #include "redisearch_api.h"
@@ -524,11 +524,6 @@ void IndexSpec_AddToInfo(RedisModuleInfoCtx *ctx, IndexSpec *sp);
  */
 size_t IndexSpec_VectorIndexSize(IndexSpec *sp);
 
-FieldSpecStats FieldSpec_GetStats(const FieldSpec *fs, IndexSpec *sp);
-
-VectorIndexStats IndexSpec_GetVectorIndexesStats(IndexSpec *sp);
-
-VectorIndexStats IndexSpec_GetVectorIndexStats(IndexSpec *sp, const FieldSpec *fs);
 /**
  * Gets the next text id from the index. This does not currently
  * modify the index
